@@ -1,23 +1,25 @@
 module github.com/projectdiscovery/subfinder/v2
 
-go 1.21
+go 1.24.0
+
+toolchain go1.24.1
 
 require (
 	github.com/corpix/uarand v0.2.0
 	github.com/hako/durafmt v0.0.0-20210316092057-3a2c319c1acd
 	github.com/json-iterator/go v1.1.12
 	github.com/lib/pq v1.10.9
-	github.com/projectdiscovery/chaos-client v0.5.1
-	github.com/projectdiscovery/dnsx v1.1.6
+	github.com/projectdiscovery/chaos-client v0.5.2
+	github.com/projectdiscovery/dnsx v1.2.2
 	github.com/projectdiscovery/fdmax v0.0.4
-	github.com/projectdiscovery/gologger v1.1.12
-	github.com/projectdiscovery/ratelimit v0.0.26
-	github.com/projectdiscovery/retryablehttp-go v1.0.44
-	github.com/projectdiscovery/utils v0.0.76
+	github.com/projectdiscovery/gologger v1.1.54
+	github.com/projectdiscovery/ratelimit v0.0.81
+	github.com/projectdiscovery/retryablehttp-go v1.0.115
+	github.com/projectdiscovery/utils v0.4.21
 	github.com/rs/xid v1.5.0
-	github.com/stretchr/testify v1.8.4
+	github.com/stretchr/testify v1.10.0
 	github.com/tomnomnom/linkheader v0.0.0-20180905144013-02ca5825eb80
-	golang.org/x/exp v0.0.0-20230420155640-133eef4313cb
+	golang.org/x/exp v0.0.0-20250106191152-7588d65b2ba8
 	gopkg.in/yaml.v3 v3.0.1
 )
 
