@@ -14,7 +14,6 @@ var (
 		"alienvault",
 		"anubis",
 		"bevigil",
-		"binaryedge",
 		"bufferover",
 		"c99",
 		"censys",
@@ -27,32 +26,43 @@ var (
 		"dnsdumpster",
 		"dnsdb",
 		"dnsrepo",
+		"domainsproject",
+		"driftnet",
 		"fofa",
 		"fullhunt",
 		"github",
 		"hackertarget",
 		"intelx",
 		"netlas",
-		"passivetotal",
+		"onyphe",
 		"quake",
+		"pugrecon",
 		"rapiddns",
 		"redhuntlabs",
 		// "riddler", // failing due to cloudfront protection
 		"robtex",
+		"rsecloud",
 		"securitytrails",
+		"profundis",
 		"shodan",
 		"sitedossier",
 		"threatbook",
+		"threatcrowd",
 		"virustotal",
 		"waybackarchive",
 		"whoisxmlapi",
+		"windvane",
 		"zoomeyeapi",
-		"hunter",
 		"leakix",
 		"facebook",
 		// "threatminer",
 		// "reconcloud",
+		"reconeer",
 		"builtwith",
+		"hudsonrock",
+		"digitalyama",
+		"merklemap",
+		"thc",
 	}
 
 	expectedDefaultSources = []string{
@@ -68,43 +78,49 @@ var (
 		"crtsh",
 		"digitorus",
 		"dnsdumpster",
+		"domainsproject",
 		"dnsrepo",
+		"driftnet",
 		"fofa",
 		"fullhunt",
 		"hackertarget",
 		"intelx",
-		"passivetotal",
+		"onyphe",
 		"quake",
 		"redhuntlabs",
 		"robtex",
 		// "riddler", // failing due to cloudfront protection
+		"rsecloud",
 		"securitytrails",
+		"profundis",
 		"shodan",
+		"windvane",
 		"virustotal",
 		"whoisxmlapi",
-		"hunter",
 		"leakix",
 		"facebook",
 		// "threatminer",
 		// "reconcloud",
+		"reconeer",
 		"builtwith",
+		"digitalyama",
+		"thc",
 	}
 
 	expectedDefaultRecursiveSources = []string{
 		"alienvault",
-		"binaryedge",
 		"bufferover",
 		"certspotter",
 		"crtsh",
-		"dnsdumpster",
 		"dnsdb",
 		"digitorus",
+		"driftnet",
 		"hackertarget",
-		"passivetotal",
 		"securitytrails",
 		"virustotal",
 		"leakix",
 		"facebook",
+		"merklemap",
 		// "reconcloud",
 	}
 )
