@@ -11,8 +11,6 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/pkg/errors"
-
 	"github.com/projectdiscovery/gologger"
 	contextutil "github.com/projectdiscovery/utils/context"
 	fileutil "github.com/projectdiscovery/utils/file"
@@ -36,15 +34,16 @@ type Runner struct {
 // the configuration options, configuring sources, reading lists
 // and setting up loggers, etc.
 func NewRunner(options *Options) (*Runner, error) {
+	options.ConfigureOutput()
 	runner := &Runner{options: options}
 
 	// Check if the application loading with any provider configuration, then take it
 	// Otherwise load the default provider config
 	if fileutil.FileExists(options.ProviderConfig) {
-		// gologger.Info().Msgf("Loading provider config from %s", options.ProviderConfig)
+		gologger.Info().Msgf("Loading provider config from %s", options.ProviderConfig)
 		options.loadProvidersFrom(options.ProviderConfig)
 	} else {
-		// gologger.Info().Msgf("Loading provider config from the default location: %s", defaultProviderConfigLocation)
+		gologger.Info().Msgf("Loading provider config from the default location: %s", defaultProviderConfigLocation)
 		options.loadProvidersFrom(defaultProviderConfigLocation)
 	}
 
@@ -95,7 +94,9 @@ func (r *Runner) RunEnumerationWithCtx(ctx context.Context) error {
 			return err
 		}
 		err = r.EnumerateMultipleDomainsWithCtx(ctx, f, outputs)
-		f.Close()
+		if closeErr := f.Close(); closeErr != nil {
+			gologger.Error().Msgf("Error closing file %s: %s", r.options.DomainsFile, closeErr)
+		}
 		return err
 	}
 
@@ -115,12 +116,14 @@ func (r *Runner) EnumerateMultipleDomains(reader io.Reader, writers []io.Writer)
 // EnumerateMultipleDomainsWithCtx enumerates subdomains for multiple domains
 // We keep enumerating subdomains for a given domain until we reach an error
 func (r *Runner) EnumerateMultipleDomainsWithCtx(ctx context.Context, reader io.Reader, writers []io.Writer) error {
+	var err error
 	scanner := bufio.NewScanner(reader)
 	ip, _ := regexp.Compile(`^([0-9\.]+$)`)
 	for scanner.Scan() {
-		domain, err := normalizeLowercase(scanner.Text())
-		isIp := ip.MatchString(domain)
-		if errors.Is(err, ErrEmptyInput) || (r.options.ExcludeIps && isIp) {
+		domain := preprocessDomain(scanner.Text())
+		domain = replacer.Replace(domain)
+
+		if domain == "" || (r.options.ExcludeIps && ip.MatchString(domain)) {
 			continue
 		}
 
@@ -136,9 +139,11 @@ func (r *Runner) EnumerateMultipleDomainsWithCtx(ctx context.Context, reader io.
 				return err
 			}
 
-			err = r.EnumerateSingleDomainWithCtx(ctx, domain, append(writers, file))
+			_, err = r.EnumerateSingleDomainWithCtx(ctx, domain, append(writers, file))
 
-			file.Close()
+			if closeErr := file.Close(); closeErr != nil {
+				gologger.Error().Msgf("Error closing file %s: %s", r.options.OutputFile, closeErr)
+			}
 		} else if r.options.OutputDirectory != "" {
 			outputFile := path.Join(r.options.OutputDirectory, domain)
 			if r.options.JSON {
@@ -154,11 +159,13 @@ func (r *Runner) EnumerateMultipleDomainsWithCtx(ctx context.Context, reader io.
 				return err
 			}
 
-			err = r.EnumerateSingleDomainWithCtx(ctx, domain, append(writers, file))
+			_, err = r.EnumerateSingleDomainWithCtx(ctx, domain, append(writers, file))
 
-			file.Close()
+			if closeErr := file.Close(); closeErr != nil {
+				gologger.Error().Msgf("Error closing file %s: %s", outputFile, closeErr)
+			}
 		} else {
-			err = r.EnumerateSingleDomainWithCtx(ctx, domain, writers)
+			_, err = r.EnumerateSingleDomainWithCtx(ctx, domain, writers)
 		}
 		if err != nil {
 			return err
