@@ -73,8 +73,8 @@ func stripRegexString(val string) string {
 	return fmt.Sprint("^", val, "$")
 }
 
-// configureOutput configures the output on the screen
-func (options *Options) configureOutput() {
+// ConfigureOutput configures the output on the screen
+func (options *Options) ConfigureOutput() {
 	// If the user desires verbose output, show verbose output
 	if options.Verbose {
 		gologger.DefaultLogger.SetMaxLevel(levels.LevelVerbose)
