@@ -11,6 +11,8 @@ import (
 	"regexp"
 	"strconv"
 	"strings"
+	"sync"
+	"sync/atomic"
 	"time"
 
 	jsoniter "github.com/json-iterator/go"
@@ -40,16 +42,18 @@ type response struct {
 type Source struct {
 	apiKeys   []string
 	timeTaken time.Duration
-	errors    int
-	results   int
+	errors    atomic.Int32
+	results   atomic.Int32
+	requests  atomic.Int32
 	skipped   bool
 }
 
 // Run function returns all subdomains found with the service
 func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Session) <-chan subscraping.Result {
 	results := make(chan subscraping.Result)
-	s.errors = 0
-	s.results = 0
+	s.errors.Store(0)
+	s.results.Store(0)
+	s.requests.Store(0)
 
 	go func() {
 		defer func(startTime time.Time) {
@@ -80,26 +84,17 @@ func (s *Source) enumerate(ctx context.Context, searchURL string, domainRegexp *
 	}
 
 	token := tokens.Get()
-
-	if token.RetryAfter > 0 {
-		if len(tokens.pool) == 1 {
-			gologger.Verbose().Label(s.Name()).Msgf("GitHub Search request rate limit exceeded, waiting for %d seconds before retry... \n", token.RetryAfter)
-			time.Sleep(time.Duration(token.RetryAfter) * time.Second)
-		} else {
-			token = tokens.Get()
-		}
-	}
-
 	headers := map[string]string{
 		"Accept": "application/vnd.github.v3.text-match+json", "Authorization": "token " + token.Hash,
 	}
 
 	// Initial request to GitHub search
+	s.requests.Add(1)
 	resp, err := session.Get(ctx, searchURL, "", headers)
 	isForbidden := resp != nil && resp.StatusCode == http.StatusForbidden
 	if err != nil && !isForbidden {
 		results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
-		s.errors++
+		s.errors.Add(1)
 		session.DiscardHTTPResponse(resp)
 		return
 	}
@@ -109,7 +104,7 @@ func (s *Source) enumerate(ctx context.Context, searchURL string, domainRegexp *
 	if isForbidden && ratelimitRemaining == 0 {
 		retryAfterSeconds, _ := strconv.ParseInt(resp.Header.Get("Retry-After"), 10, 64)
 		tokens.setCurrentTokenExceeded(retryAfterSeconds)
-		resp.Body.Close()
+		session.DiscardHTTPResponse(resp)
 
 		s.enumerate(ctx, searchURL, domainRegexp, tokens, session, results)
 	}
@@ -120,17 +115,17 @@ func (s *Source) enumerate(ctx context.Context, searchURL string, domainRegexp *
 	err = jsoniter.NewDecoder(resp.Body).Decode(&data)
 	if err != nil {
 		results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
-		s.errors++
-		resp.Body.Close()
+		s.errors.Add(1)
+		session.DiscardHTTPResponse(resp)
 		return
 	}
 
-	resp.Body.Close()
+	session.DiscardHTTPResponse(resp)
 
 	err = s.proccesItems(ctx, data.Items, domainRegexp, s.Name(), session, results)
 	if err != nil {
 		results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
-		s.errors++
+		s.errors.Add(1)
 		return
 	}
 
@@ -138,11 +133,16 @@ func (s *Source) enumerate(ctx context.Context, searchURL string, domainRegexp *
 	linksHeader := linkheader.Parse(resp.Header.Get("Link"))
 	// Process the next link recursively
 	for _, link := range linksHeader {
+		select {
+		case <-ctx.Done():
+			return
+		default:
+		}
 		if link.Rel == "next" {
 			nextURL, err := url.QueryUnescape(link.URL)
 			if err != nil {
 				results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
-				s.errors++
+				s.errors.Add(1)
 				return
 			}
 			s.enumerate(ctx, nextURL, domainRegexp, tokens, session, results)
@@ -150,42 +150,90 @@ func (s *Source) enumerate(ctx context.Context, searchURL string, domainRegexp *
 	}
 }
 
-// proccesItems procceses github response items
+// proccesItems process github response items
 func (s *Source) proccesItems(ctx context.Context, items []item, domainRegexp *regexp.Regexp, name string, session *subscraping.Session, results chan subscraping.Result) error {
-	for _, item := range items {
-		// find subdomains in code
-		resp, err := session.SimpleGet(ctx, rawURL(item.HTMLURL))
-		if err != nil {
-			if resp != nil && resp.StatusCode != http.StatusNotFound {
+	var wg sync.WaitGroup
+	errChan := make(chan error, len(items))
+
+	for _, responseItem := range items {
+		select {
+		case <-ctx.Done():
+			return ctx.Err()
+		default:
+		}
+		wg.Add(1)
+		go func(responseItem item) {
+			defer wg.Done()
+
+			select {
+			case <-ctx.Done():
+				return
+			default:
+			}
+
+			s.requests.Add(1)
+			resp, err := session.SimpleGet(ctx, rawURL(responseItem.HTMLURL))
+			if err != nil {
+				if resp != nil && resp.StatusCode != http.StatusNotFound {
+					session.DiscardHTTPResponse(resp)
+				}
+				errChan <- err
+				return
+			}
+
+			if resp.StatusCode == http.StatusOK {
+				scanner := bufio.NewScanner(resp.Body)
+				for scanner.Scan() {
+					select {
+					case <-ctx.Done():
+						session.DiscardHTTPResponse(resp)
+						return
+					default:
+					}
+					line := scanner.Text()
+					if line == "" {
+						continue
+					}
+					for _, subdomain := range domainRegexp.FindAllString(normalizeContent(line), -1) {
+						select {
+						case <-ctx.Done():
+							session.DiscardHTTPResponse(resp)
+							return
+						case results <- subscraping.Result{Source: name, Type: subscraping.Subdomain, Value: subdomain}:
+							s.results.Add(1)
+						}
+					}
+				}
 				session.DiscardHTTPResponse(resp)
 			}
+
+			for _, textMatch := range responseItem.TextMatches {
+				select {
+				case <-ctx.Done():
+					return
+				default:
+				}
+				for _, subdomain := range domainRegexp.FindAllString(normalizeContent(textMatch.Fragment), -1) {
+					select {
+					case <-ctx.Done():
+						return
+					case results <- subscraping.Result{Source: name, Type: subscraping.Subdomain, Value: subdomain}:
+						s.results.Add(1)
+					}
+				}
+			}
+		}(responseItem)
+	}
+
+	wg.Wait()
+	close(errChan)
+
+	for err := range errChan {
+		if err != nil {
 			return err
 		}
-
-		if resp.StatusCode == http.StatusOK {
-			scanner := bufio.NewScanner(resp.Body)
-			for scanner.Scan() {
-				line := scanner.Text()
-				if line == "" {
-					continue
-				}
-				for _, subdomain := range domainRegexp.FindAllString(normalizeContent(line), -1) {
-					results <- subscraping.Result{Source: name, Type: subscraping.Subdomain, Value: subdomain}
-					s.results++
-
-				}
-			}
-			resp.Body.Close()
-		}
-
-		// find subdomains in text matches
-		for _, textMatch := range item.TextMatches {
-			for _, subdomain := range domainRegexp.FindAllString(normalizeContent(textMatch.Fragment), -1) {
-				results <- subscraping.Result{Source: name, Type: subscraping.Subdomain, Value: subdomain}
-				s.results++
-			}
-		}
 	}
+
 	return nil
 }
 
@@ -222,8 +270,12 @@ func (s *Source) HasRecursiveSupport() bool {
 	return false
 }
 
+func (s *Source) KeyRequirement() subscraping.KeyRequirement {
+	return subscraping.RequiredKey
+}
+
 func (s *Source) NeedsKey() bool {
-	return true
+	return s.KeyRequirement() == subscraping.RequiredKey
 }
 
 func (s *Source) AddApiKeys(keys []string) {
@@ -232,8 +284,9 @@ func (s *Source) AddApiKeys(keys []string) {
 
 func (s *Source) Statistics() subscraping.Statistics {
 	return subscraping.Statistics{
-		Errors:    s.errors,
-		Results:   s.results,
+		Errors:    int(s.errors.Load()),
+		Results:   int(s.results.Load()),
+		Requests:  int(s.requests.Load()),
 		TimeTaken: s.timeTaken,
 		Skipped:   s.skipped,
 	}
