@@ -9,6 +9,7 @@ import (
 	"regexp"
 	"strings"
 	"sync"
+	"sync/atomic"
 	"time"
 
 	jsoniter "github.com/json-iterator/go"
@@ -18,7 +19,12 @@ import (
 
 // Source is the passive scraping agent
 type Source struct {
-	apiKeys []string
+	apiKeys   []string
+	timeTaken time.Duration
+	errors    atomic.Int32
+	results   atomic.Int32
+	requests  atomic.Int32
+	skipped   bool
 }
 
 type item struct {
@@ -31,9 +37,15 @@ type item struct {
 // Run function returns all subdomains found with the service
 func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Session) <-chan subscraping.Result {
 	results := make(chan subscraping.Result)
+	s.errors.Store(0)
+	s.results.Store(0)
+	s.requests.Store(0)
 
 	go func() {
-		defer close(results)
+		defer func(startTime time.Time) {
+			s.timeTaken = time.Since(startTime)
+			close(results)
+		}(time.Now())
 
 		randomApiKey := subscraping.PickRandom(s.apiKeys, s.Name())
 		if randomApiKey == "" {
@@ -57,19 +69,22 @@ func (s *Source) enumerate(ctx context.Context, searchURL string, domainRegexp *
 	default:
 	}
 
+	s.requests.Add(1)
 	resp, err := session.Get(ctx, searchURL, "", headers)
 	if err != nil && resp == nil {
 		results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+		s.errors.Add(1)
 		session.DiscardHTTPResponse(resp)
 		return
 	}
 
-	defer resp.Body.Close()
+	defer session.DiscardHTTPResponse(resp)
 
 	var items []item
 	err = jsoniter.NewDecoder(resp.Body).Decode(&items)
 	if err != nil {
 		results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+		s.errors.Add(1)
 		return
 	}
 
@@ -80,12 +95,14 @@ func (s *Source) enumerate(ctx context.Context, searchURL string, domainRegexp *
 		go func(item item) {
 			// The original item.Path causes 404 error because the Gitlab API is expecting the url encoded path
 			fileUrl := fmt.Sprintf("https://gitlab.com/api/v4/projects/%d/repository/files/%s/raw?ref=%s", item.ProjectId, url.QueryEscape(item.Path), item.Ref)
+			s.requests.Add(1)
 			resp, err := session.Get(ctx, fileUrl, "", headers)
 			if err != nil {
 				if resp == nil || (resp != nil && resp.StatusCode != http.StatusNotFound) {
 					session.DiscardHTTPResponse(resp)
 
 					results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+					s.errors.Add(1)
 					return
 				}
 			}
@@ -99,27 +116,29 @@ func (s *Source) enumerate(ctx context.Context, searchURL string, domainRegexp *
 					}
 					for _, subdomain := range domainRegexp.FindAllString(line, -1) {
 						results <- subscraping.Result{Source: s.Name(), Type: subscraping.Subdomain, Value: subdomain}
+						s.results.Add(1)
 					}
 				}
-				resp.Body.Close()
+				session.DiscardHTTPResponse(resp)
 			}
 			defer wg.Done()
 		}(it)
 	}
 
-	// Links header, first, next, last...
 	linksHeader := linkheader.Parse(resp.Header.Get("Link"))
-	// Process the next link recursively
 	for _, link := range linksHeader {
+		select {
+		case <-ctx.Done():
+			return
+		default:
+		}
 		if link.Rel == "next" {
 			nextURL, err := url.QueryUnescape(link.URL)
 			if err != nil {
 				results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+				s.errors.Add(1)
 				return
 			}
-
-			// TODO: hardcoded for testing, should be a source internal rate limit #718
-			time.Sleep(2 * time.Second)
 
 			s.enumerate(ctx, nextURL, domainRegexp, headers, session, results)
 		}
@@ -146,10 +165,25 @@ func (s *Source) HasRecursiveSupport() bool {
 	return false
 }
 
+func (s *Source) KeyRequirement() subscraping.KeyRequirement {
+	return subscraping.RequiredKey
+}
+
 func (s *Source) NeedsKey() bool {
-	return true
+	return s.KeyRequirement() == subscraping.RequiredKey
 }
 
 func (s *Source) AddApiKeys(keys []string) {
 	s.apiKeys = keys
+}
+
+// Statistics returns the statistics for the source
+func (s *Source) Statistics() subscraping.Statistics {
+	return subscraping.Statistics{
+		Errors:    int(s.errors.Load()),
+		Results:   int(s.results.Load()),
+		Requests:  int(s.requests.Load()),
+		TimeTaken: s.timeTaken,
+		Skipped:   s.skipped,
+	}
 }
