@@ -3,17 +3,23 @@ package rapiddns
 
 import (
 	"context"
+	"fmt"
 	"io"
+	"regexp"
+	"strconv"
 	"time"
 
 	"github.com/projectdiscovery/subfinder/v2/pkg/subscraping"
 )
+
+var pagePattern = regexp.MustCompile(`class="page-link" href="/subdomain/[^"]+\?page=(\d+)">`)
 
 // Source is the passive scraping agent
 type Source struct {
 	timeTaken time.Duration
 	errors    int
 	results   int
+	requests  int
 }
 
 // Run function returns all subdomains found with the service
@@ -21,6 +27,7 @@ func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Se
 	results := make(chan subscraping.Result)
 	s.errors = 0
 	s.results = 0
+	s.requests = 0
 
 	go func() {
 		defer func(startTime time.Time) {
@@ -28,28 +35,57 @@ func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Se
 			close(results)
 		}(time.Now())
 
-		resp, err := session.SimpleGet(ctx, "https://rapiddns.io/subdomain/"+domain+"?full=1")
-		if err != nil {
-			results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
-			s.errors++
+		page := 1
+		maxPages := 1
+		for {
+			select {
+			case <-ctx.Done():
+				return
+			default:
+			}
+			s.requests++
+			resp, err := session.SimpleGet(ctx, fmt.Sprintf("https://rapiddns.io/subdomain/%s?page=%d&full=1", domain, page))
+			if err != nil {
+				results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+				s.errors++
+				session.DiscardHTTPResponse(resp)
+				return
+			}
+
+			body, err := io.ReadAll(resp.Body)
+			if err != nil {
+				results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+				s.errors++
+				session.DiscardHTTPResponse(resp)
+				return
+			}
+
 			session.DiscardHTTPResponse(resp)
-			return
-		}
 
-		body, err := io.ReadAll(resp.Body)
-		if err != nil {
-			results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
-			s.errors++
-			resp.Body.Close()
-			return
-		}
+			src := string(body)
+			for _, subdomain := range session.Extractor.Extract(src) {
+				select {
+				case <-ctx.Done():
+					return
+				case results <- subscraping.Result{Source: s.Name(), Type: subscraping.Subdomain, Value: subdomain}:
+					s.results++
+				}
+			}
 
-		resp.Body.Close()
+			if maxPages == 1 {
+				matches := pagePattern.FindAllStringSubmatch(src, -1)
+				if len(matches) > 0 {
+					lastMatch := matches[len(matches)-1]
+					if len(lastMatch) > 1 {
+						maxPages, _ = strconv.Atoi(lastMatch[1])
+					}
+				}
+			}
 
-		src := string(body)
-		for _, subdomain := range session.Extractor.Extract(src) {
-			results <- subscraping.Result{Source: s.Name(), Type: subscraping.Subdomain, Value: subdomain}
-			s.results++
+			if page >= maxPages {
+				break
+			}
+			page++
 		}
 	}()
 
@@ -69,8 +105,12 @@ func (s *Source) HasRecursiveSupport() bool {
 	return false
 }
 
+func (s *Source) KeyRequirement() subscraping.KeyRequirement {
+	return subscraping.NoKey
+}
+
 func (s *Source) NeedsKey() bool {
-	return false
+	return s.KeyRequirement() == subscraping.RequiredKey
 }
 
 func (s *Source) AddApiKeys(_ []string) {
@@ -82,5 +122,6 @@ func (s *Source) Statistics() subscraping.Statistics {
 		Errors:    s.errors,
 		Results:   s.results,
 		TimeTaken: s.timeTaken,
+		Requests:  s.requests,
 	}
 }
