@@ -17,6 +17,7 @@ type Source struct {
 	timeTaken time.Duration
 	errors    int
 	results   int
+	requests  int
 	skipped   bool
 }
 
@@ -25,6 +26,7 @@ type dnsdbLookupResponse struct {
 	Subdomains []string `json:"subdomains"`
 	Result     int      `json:"result"`
 	Error      string   `json:"error"`
+	More       bool     `json:"more"`
 }
 
 // Run function returns all subdomains found with the service
@@ -32,6 +34,7 @@ func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Se
 	results := make(chan subscraping.Result)
 	s.errors = 0
 	s.results = 0
+	s.requests = 0
 
 	go func() {
 		defer func(startTime time.Time) {
@@ -45,36 +48,57 @@ func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Se
 			return
 		}
 
-		searchURL := fmt.Sprintf("https://api.shodan.io/dns/domain/%s?key=%s", domain, randomApiKey)
-		resp, err := session.SimpleGet(ctx, searchURL)
-		if err != nil {
-			session.DiscardHTTPResponse(resp)
-			return
-		}
-
-		defer resp.Body.Close()
-
-		var response dnsdbLookupResponse
-		err = jsoniter.NewDecoder(resp.Body).Decode(&response)
-		if err != nil {
-			results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
-			s.errors++
-			return
-		}
-
-		if response.Error != "" {
-			results <- subscraping.Result{
-				Source: s.Name(), Type: subscraping.Error, Error: fmt.Errorf("%v", response.Error),
+		page := 1
+		for {
+			select {
+			case <-ctx.Done():
+				return
+			default:
 			}
-			s.errors++
-			return
-		}
 
-		for _, data := range response.Subdomains {
-			results <- subscraping.Result{
-				Source: s.Name(), Type: subscraping.Subdomain, Value: fmt.Sprintf("%s.%s", data, domain),
+			searchURL := fmt.Sprintf("https://api.shodan.io/dns/domain/%s?key=%s&page=%d", domain, randomApiKey, page)
+			s.requests++
+			resp, err := session.SimpleGet(ctx, searchURL)
+			if err != nil {
+				session.DiscardHTTPResponse(resp)
+				return
 			}
-			s.results++
+
+			defer session.DiscardHTTPResponse(resp)
+
+			var response dnsdbLookupResponse
+			err = jsoniter.NewDecoder(resp.Body).Decode(&response)
+			if err != nil {
+				results <- subscraping.Result{Source: s.Name(), Type: subscraping.Error, Error: err}
+				s.errors++
+				return
+			}
+
+			if response.Error != "" {
+				results <- subscraping.Result{
+					Source: s.Name(), Type: subscraping.Error, Error: fmt.Errorf("%v", response.Error),
+				}
+				s.errors++
+				return
+			}
+
+			for _, data := range response.Subdomains {
+				select {
+				case <-ctx.Done():
+					return
+				default:
+				}
+				value := fmt.Sprintf("%s.%s", data, response.Domain)
+				results <- subscraping.Result{
+					Source: s.Name(), Type: subscraping.Subdomain, Value: value,
+				}
+				s.results++
+			}
+
+			if !response.More {
+				break
+			}
+			page++
 		}
 	}()
 
@@ -94,8 +118,12 @@ func (s *Source) HasRecursiveSupport() bool {
 	return false
 }
 
+func (s *Source) KeyRequirement() subscraping.KeyRequirement {
+	return subscraping.RequiredKey
+}
+
 func (s *Source) NeedsKey() bool {
-	return true
+	return s.KeyRequirement() == subscraping.RequiredKey
 }
 
 func (s *Source) AddApiKeys(keys []string) {
@@ -106,6 +134,7 @@ func (s *Source) Statistics() subscraping.Statistics {
 	return subscraping.Statistics{
 		Errors:    s.errors,
 		Results:   s.results,
+		Requests:  s.requests,
 		TimeTaken: s.timeTaken,
 		Skipped:   s.skipped,
 	}
