@@ -5,7 +5,6 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"math/rand"
 	"net/http"
 	"regexp"
 	"time"
@@ -19,55 +18,12 @@ const SleepRandIntn = 5
 
 var reNext = regexp.MustCompile(`<a href="([A-Za-z0-9/.]+)"><b>`)
 
-type agent struct {
-	results chan subscraping.Result
-	errors  int
-	session *subscraping.Session
-}
-
-func (a *agent) enumerate(ctx context.Context, baseURL string) {
-	select {
-	case <-ctx.Done():
-		return
-	default:
-	}
-
-	resp, err := a.session.SimpleGet(ctx, baseURL)
-	isnotfound := resp != nil && resp.StatusCode == http.StatusNotFound
-	if err != nil && !isnotfound {
-		a.results <- subscraping.Result{Source: "sitedossier", Type: subscraping.Error, Error: err}
-		a.errors++
-		a.session.DiscardHTTPResponse(resp)
-		return
-	}
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		a.results <- subscraping.Result{Source: "sitedossier", Type: subscraping.Error, Error: err}
-		a.errors++
-		resp.Body.Close()
-		return
-	}
-	resp.Body.Close()
-
-	src := string(body)
-	for _, match := range a.session.Extractor.Extract(src) {
-		a.results <- subscraping.Result{Source: "sitedossier", Type: subscraping.Subdomain, Value: match}
-	}
-
-	match1 := reNext.FindStringSubmatch(src)
-	time.Sleep(time.Duration((3 + rand.Intn(SleepRandIntn))) * time.Second)
-
-	if len(match1) > 0 {
-		a.enumerate(ctx, "http://www.sitedossier.com"+match1[1])
-	}
-}
-
 // Source is the passive scraping agent
 type Source struct {
 	timeTaken time.Duration
 	errors    int
 	results   int
+	requests  int
 }
 
 // Run function returns all subdomains found with the service
@@ -75,24 +31,60 @@ func (s *Source) Run(ctx context.Context, domain string, session *subscraping.Se
 	results := make(chan subscraping.Result)
 	s.errors = 0
 	s.results = 0
-
-	a := agent{
-		session: session,
-		results: results,
-	}
+	s.requests = 0
 
 	go func() {
 		defer func(startTime time.Time) {
 			s.timeTaken = time.Since(startTime)
-			close(a.results)
+			close(results)
 		}(time.Now())
 
-		a.enumerate(ctx, fmt.Sprintf("http://www.sitedossier.com/parentdomain/%s", domain))
-		s.errors = a.errors
-		s.results = len(a.results)
+		s.enumerate(ctx, session, fmt.Sprintf("http://www.sitedossier.com/parentdomain/%s", domain), results)
 	}()
 
-	return a.results
+	return results
+}
+
+func (s *Source) enumerate(ctx context.Context, session *subscraping.Session, baseURL string, results chan subscraping.Result) {
+	select {
+	case <-ctx.Done():
+		return
+	default:
+	}
+
+	s.requests++
+	resp, err := session.SimpleGet(ctx, baseURL)
+	isnotfound := resp != nil && resp.StatusCode == http.StatusNotFound
+	if err != nil && !isnotfound {
+		results <- subscraping.Result{Source: "sitedossier", Type: subscraping.Error, Error: err}
+		s.errors++
+		session.DiscardHTTPResponse(resp)
+		return
+	}
+
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		results <- subscraping.Result{Source: "sitedossier", Type: subscraping.Error, Error: err}
+		s.errors++
+		session.DiscardHTTPResponse(resp)
+		return
+	}
+	session.DiscardHTTPResponse(resp)
+
+	src := string(body)
+	for _, subdomain := range session.Extractor.Extract(src) {
+		select {
+		case <-ctx.Done():
+			return
+		case results <- subscraping.Result{Source: "sitedossier", Type: subscraping.Subdomain, Value: subdomain}:
+			s.results++
+		}
+	}
+
+	match := reNext.FindStringSubmatch(src)
+	if len(match) > 0 {
+		s.enumerate(ctx, session, fmt.Sprintf("http://www.sitedossier.com%s", match[1]), results)
+	}
 }
 
 // Name returns the name of the source
@@ -108,8 +100,12 @@ func (s *Source) HasRecursiveSupport() bool {
 	return false
 }
 
+func (s *Source) KeyRequirement() subscraping.KeyRequirement {
+	return subscraping.NoKey
+}
+
 func (s *Source) NeedsKey() bool {
-	return false
+	return s.KeyRequirement() == subscraping.RequiredKey
 }
 
 func (s *Source) AddApiKeys(_ []string) {
@@ -121,5 +117,6 @@ func (s *Source) Statistics() subscraping.Statistics {
 		Errors:    s.errors,
 		Results:   s.results,
 		TimeTaken: s.timeTaken,
+		Requests:  s.requests,
 	}
 }
